+++ packages/commonwealth/client/scripts/views/components/Polls/DeletePollModal/DeletePollModal.scss
@@ -1,0 +1,9 @@
+@import '../../../../../styles/shared';
+
+.DeletePollModal {
+  width: 100%;
+
+  @include smallInclusive {
+    width: 100%;
+  }
+}
