+++ packages/commonwealth/client/scripts/views/components/Polls/PollCard/PollCard.scss
@@ -1,0 +1,45 @@
+@import '../../../../../styles/shared';
+
+.PollCard.Card {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  padding: 0;
+  width: 248px;
+
+  @include mediumSmallInclusive {
+    width: 100%;
+  }
+
+  .poll-title-section {
+    width: 100%;
+    border-bottom: 1px solid $neutral-200;
+    min-height: 32px;
+    padding: 16px;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+
+    .poll-delete-button {
+      cursor: pointer;
+    }
+
+    .poll-title-text.Text {
+      width: 100%;
+    }
+  }
+
+  .poll-voting-section {
+    border-bottom: 1px solid $neutral-200;
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+    padding: 12px 16px 12px 16px;
+    width: 100%;
+
+    &:empty {
+      display: none;
+    }
+  }
+}
