+++ packages/commonwealth/client/scripts/views/components/Polls/PollOptions/PollOptions.scss
@@ -1,0 +1,7 @@
+@import '../../../../../styles/shared';
+
+.PollOptions {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
