+++ packages/commonwealth/client/scripts/views/components/Polls/ResultsSections/ResultsSections.scss
@@ -1,0 +1,37 @@
+@import '../../../../../styles/shared';
+
+.ResultsSections {
+  display: flex;
+  flex-direction: column;
+  padding: 16px;
+  width: 100%;
+
+  .results-header {
+    align-items: center;
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 16px;
+    width: 100%;
+
+    .clickable {
+      cursor: pointer;
+      color: $primary-300;
+    }
+  }
+
+  .results-content {
+    display: flex;
+    flex-direction: column;
+    gap: 24px;
+  }
+
+  .more-options.Text {
+    color: $neutral-500;
+    justify-content: flex-end;
+    margin-top: 8px;
+  }
+
+  .button-icon.Icon {
+    color: $green-500;
+  }
+}
