+++ packages/commonwealth/client/scripts/views/components/Polls/index.ts
@@ -1,0 +1,29 @@
+import type { CastVoteProps } from './CastVoteSection/CastVoteSection';
+import { CastVoteSection } from './CastVoteSection/CastVoteSection';
+import { DeletePollModal } from './DeletePollModal/DeletePollModal';
+import type { PollCardProps } from './PollCard/PollCard';
+import { PollCard } from './PollCard/PollCard';
+import type {
+  PollOptionProps,
+  VoteInformation,
+} from './PollOptions/PollOptions';
+import { PollOptions } from './PollOptions/PollOptions';
+import type { ResultsSectionProps } from './ResultsSections/ResultsSections';
+import { ResultsSections } from './ResultsSections/ResultsSections';
+import { VoteDisplay } from './VoteDisplay/VoteDisplay';
+export {
+  CastVoteSection,
+  DeletePollModal,
+  PollCard,
+  PollOptions,
+  ResultsSections,
+  VoteDisplay,
+};
+
+export type {
+  CastVoteProps,
+  PollCardProps,
+  PollOptionProps,
+  ResultsSectionProps,
+  VoteInformation,
+};
