+++ packages/commonwealth/client/scripts/views/components/Polls/utils.ts
@@ -1,0 +1,7 @@
+export const buildVoteDirectionString = (voteOption: string) => {
+  if (!voteOption) {
+    return '';
+  }
+
+  return `You voted "${voteOption}"`;
+};
