--- packages/commonwealth/client/styles/components/poll_card.scss
@@ -1,142 +1,0 @@
-@import '../shared';
-
-.PollOptions {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-.CastVoteSection {
-  align-items: center;
-  display: flex;
-
-  .time-remaining-text.Text {
-    color: $neutral-500;
-    margin-left: 16px;
-  }
-}
-
-.VoteDisplay {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  justify-content: center;
-  width: 100%;
-
-  .vote-direction {
-    display: flex;
-
-    .vote-check-icon.Icon {
-      color: $green-500;
-      margin-right: 8px;
-    }
-  }
-
-  .completed-vote-information {
-    align-items: center;
-    background: $green-50;
-    border-radius: $border-radius-corners;
-    display: flex;
-    flex-direction: column;
-    padding: 8px 4px 8px 4px;
-    width: 100%;
-    text-align: center;
-
-    .direction-text {
-      color: $neutral-500;
-    }
-  }
-
-  .time-remaining-text.Text {
-    color: $neutral-500;
-  }
-}
-
-.ResultsSection {
-  display: flex;
-  flex-direction: column;
-  padding: 16px;
-  width: 100%;
-
-  .results-header {
-    align-items: center;
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 16px;
-    width: 100%;
-
-    .clickable {
-      cursor: pointer;
-      color: $primary-300;
-    }
-  }
-
-  .results-content {
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
-  }
-
-  .more-options.Text {
-    color: $neutral-500;
-    justify-content: flex-end;
-    margin-top: 8px;
-  }
-
-  .button-icon.Icon {
-    color: $green-500;
-  }
-}
-
-.PollCard.Card {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  padding: 0;
-  width: 248px;
-
-  @include mediumSmallInclusive {
-    width: 100%;
-  }
-
-  .poll-title-section {
-    width: 100%;
-    border-bottom: 1px solid $neutral-200;
-    min-height: 32px;
-    padding: 16px;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-
-    .poll-delete-button {
-      cursor: pointer;
-    }
-
-    .poll-title-text.Text {
-      width: 100%;
-    }
-  }
-
-  .poll-voting-section {
-    border-bottom: 1px solid $neutral-200;
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-    padding: 12px 16px 12px 16px;
-    width: 100%;
-
-    &:empty {
-      display: none;
-    }
-  }
-}
-
-.DeleteThreadModal {
-  width: 100%;
-
-  @include smallInclusive {
-    width: 100%;
-  }
-}
